+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.8.0
+
+* Add Type#unconstrained that returns a super type with all user specific
+  constraints removed on sub types, recursively on all non scalar types.
+
+* Add high-order types, such as `Collection<T> = [T]`
+
 # 0.7.0 / 2019-02-28
 
 * Implement (basic) @import feature, working with relative paths